refactor(neon_files): Refactor package structure
