+++ packages/neon/neon_files/lib/src/dialogs/create_folder.dart
@@ -1,4 +1,7 @@
-part of '../neon_files.dart';
+import 'package:flutter/material.dart';
+import 'package:neon/utils.dart';
+import 'package:neon/widgets.dart';
+import 'package:neon_files/l10n/localizations.dart';
 
 class FilesCreateFolderDialog extends StatefulWidget {
   const FilesCreateFolderDialog({
