+++ packages/neon/neon_files/lib/src/pages/main.dart
@@ -1,4 +1,10 @@
-part of '../neon_files.dart';
+import 'package:flutter/material.dart';
+import 'package:neon/utils.dart';
+import 'package:neon/widgets.dart';
+import 'package:neon_files/l10n/localizations.dart';
+import 'package:neon_files/src/blocs/files.dart';
+import 'package:neon_files/src/dialogs/choose_create.dart';
+import 'package:neon_files/src/widgets/browser_view.dart';
 
 class FilesMainPage extends StatefulWidget {
   const FilesMainPage({
