+++ packages/neon/neon_files/lib/src/routes.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'package:neon/utils.dart';
-import 'package:neon_files/neon_files.dart';
+import 'package:neon_files/src/pages/main.dart';
 import 'package:nextcloud/nextcloud.dart';
 
 part 'routes.g.dart';
