+++ packages/neon/neon_files/pubspec.yaml
@@ -22,6 +22,7 @@ dependencies:
   go_router: ^12.0.0
   image_picker: ^1.0.0
   intl: ^0.18.0
+  meta: ^1.0.0
   neon:
     git:
       url: https://github.com/nextcloud/neon
